Time: 107 ms (82.14%), Space: 68.9 MB (91.76%) - LeetHub

+++ 1482-minimum-number-of-days-to-make-m-bouquets/1482-minimum-number-of-days-to-make-m-bouquets.cpp
@@ -21,7 +21,7 @@ private:
 public:
     int minDays(vector<int>& bloomDay, int m, int k)
     {
-        long minSize = (long)m*(long)k;
+        long minSize = (long)m*k;
         long gardenSize = bloomDay.size();
         if( gardenSize < minSize)
             return -1;
